rename conference package feature_conference
