+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Network_apps" >
         <activity
-            android:name=".conference.presentation.ConferenceActivity"
+            android:name=".feature_conference.presentation.ConferenceActivity"
             android:exported="false" />
         <activity
             android:name=".feature_basic.presentation.BasicActivity"
