--- app/src/main/java/com/survivalcoding/network_apps/conference/domain/repository/ConferenceRepository.kt
@@ -1,7 +1,0 @@
-package com.survivalcoding.network_apps.conference.domain.repository
-
-import com.survivalcoding.network_apps.conference.domain.model.ConferenceItem
-
-interface ConferenceRepository {
-    suspend fun getData(): List<ConferenceItem>?
-}
