+++ app/src/main/java/com/survivalcoding/network_apps/feature_conference/data/datasource/ConferenceURL.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.network_apps.conference.data.datasource
+package com.survivalcoding.network_apps.feature_conference.data.datasource
 
 class ConferenceURL {
     companion object {
