+++ app/src/main/java/com/survivalcoding/network_apps/feature_conference/domain/repository/ConferenceRepository.kt
@@ -1,0 +1,7 @@
+package com.survivalcoding.network_apps.feature_conference.domain.repository
+
+import com.survivalcoding.network_apps.feature_conference.domain.model.ConferenceItem
+
+interface ConferenceRepository {
+    suspend fun getData(): List<ConferenceItem>?
+}
