+++ app/src/main/java/com/survivalcoding/network_apps/feature_conference/presentation/ConferenceFragment.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.network_apps.conference.presentation
+package com.survivalcoding.network_apps.feature_conference.presentation
 
 import android.content.Intent
 import android.net.Uri
@@ -9,7 +9,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.survivalcoding.network_apps.databinding.ConferenceFragmentBinding
-import com.survivalcoding.network_apps.conference.presentation.adapter.ConferenceListAdapter
+import com.survivalcoding.network_apps.feature_conference.presentation.adapter.ConferenceListAdapter
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
