+++ app/src/main/java/com/survivalcoding/network_apps/feature_conference/presentation/ConferenceViewModel.kt
@@ -1,8 +1,8 @@
-package com.survivalcoding.network_apps.conference.presentation
+package com.survivalcoding.network_apps.feature_conference.presentation
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.liveData
-import com.survivalcoding.network_apps.conference.domain.repository.ConferenceRepository
+import com.survivalcoding.network_apps.feature_conference.domain.repository.ConferenceRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
