+++ app/src/main/java/com/survivalcoding/network_apps/feature_conference/presentation/adapter/ConferenceDiffUtilCallback.kt
@@ -1,7 +1,7 @@
-package com.survivalcoding.network_apps.conference.presentation.adapter
+package com.survivalcoding.network_apps.feature_conference.presentation.adapter
 
 import androidx.recyclerview.widget.DiffUtil
-import com.survivalcoding.network_apps.conference.domain.model.ConferenceItem
+import com.survivalcoding.network_apps.feature_conference.domain.model.ConferenceItem
 
 object ConferenceDiffUtilCallback : DiffUtil.ItemCallback<ConferenceItem>() {
     override fun areItemsTheSame(oldItem: ConferenceItem, newItem: ConferenceItem) =
