+++ app/src/main/java/com/survivalcoding/network_apps/feature_conference/presentation/adapter/ConferenceListAdapter.kt
@@ -1,10 +1,10 @@
-package com.survivalcoding.network_apps.conference.presentation.adapter
+package com.survivalcoding.network_apps.feature_conference.presentation.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import com.survivalcoding.network_apps.databinding.ConferenceListItemBinding
-import com.survivalcoding.network_apps.conference.domain.model.ConferenceItem
+import com.survivalcoding.network_apps.feature_conference.domain.model.ConferenceItem
 
 class ConferenceListAdapter(private val onLinkClicked: (String) -> Any) :
     ListAdapter<ConferenceItem, ConferenceItemViewHolder>(ConferenceDiffUtilCallback) {
