+++ app/src/main/res/layout/activity_conference.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".conference.presentation.ConferenceActivity">
+    tools:context=".feature_conference.presentation.ConferenceActivity">
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container_view"
-        android:name="com.survivalcoding.network_apps.conference.presentation.ConferenceFragment"
+        android:name="com.survivalcoding.network_apps.feature_conference.presentation.ConferenceFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
