+++ app/src/main/res/layout/conference_fragment.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".conference.presentation.ConferenceFragment">
+    tools:context=".feature_conference.presentation.ConferenceFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewConferences"
